One to many logic moved to separate file.
